republic/requirements.txt
# Fulcrum Virtual Fleet Dependencies

# Exchange Integration
ccxt>=4.0.0

# Social Media / Oracle Scanning
tweepy>=4.14.0
praw>=7.7.0

# HTTP Requests
requests>=2.31.0

# AI / NLP (The "Muscle")
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Message Bus (Nervous System)
redis>=5.0.0

# Database
# (sqlite3 is built into Python)

# Utilities
numpy>=1.26.0

# Testing
pytest
google-generativeai>=0.7.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
wheel
pandas>=2.1.0
kagglehub>=0.1.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
